requirements.txt
debugpy
flask==2.3.1
pillow
eclipse-zenoh
vaal
numpy
deepview-rt
edgefirst-schemas==1.1.1